requirements.txt
python-vlc
mediapipe
opencv-python
tensorflow